Moved root.go into main.go.
