--- cmd/root.go
@@ -1,60 +1,0 @@
-package cmd
-
-import (
-	"github.com/microsoft/abstrakt/internal/tools/logger"
-	homedir "github.com/mitchellh/go-homedir"
-	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"os"
-)
-
-var cfgFile string
-
-// rootCmd represents the base command when called without any subcommands
-
-var rootCmd = DefaultRootCommand()
-
-// Execute adds all child commands to the root command and sets flags appropriately.
-// This is called by main.main(). It only needs to happen once to the rootCmd.
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		logger.Error(err)
-		os.Exit(1)
-	}
-}
-
-func init() {
-	cobra.OnInitialize(initConfig)
-
-	// Here you will define your flags and configuration settings.
-	// Cobra supports persistent flags, which, if defined here,
-	// will be global for your application.
-
-	rootCmd.PersistentFlags().BoolP("verbose", "v", false, "Use verbose output logs")
-}
-
-// initConfig reads in config file and ENV variables if set.
-func initConfig() {
-	if cfgFile != "" {
-		// Use config file from the flag.
-		viper.SetConfigFile(cfgFile)
-	} else {
-		// Find home directory.
-		home, err := homedir.Dir()
-		if err != nil {
-			logger.Error(err)
-			os.Exit(1)
-		}
-
-		// Search config in home directory with name ".abstrakt" (without extension).
-		viper.AddConfigPath(home)
-		viper.SetConfigName(".abstrakt")
-	}
-
-	viper.AutomaticEnv() // read in environment variables that match
-
-	// If a config file is found, read it in.
-	if err := viper.ReadInConfig(); err == nil {
-		logger.Debug("Using config file:", viper.ConfigFileUsed())
-	}
-}
